Move PostgresDB to different package
