+++ src/main/java/org/commcare/formplayer/services/FormplayerStorageFactory.java
@@ -2,7 +2,7 @@ package org.commcare.formplayer.services;
 
 import org.commcare.formplayer.beans.InstallRequestBean;
 import org.commcare.formplayer.objects.SerializableFormSession;
-import org.commcare.formplayer.sandbox.PostgresDB;
+import org.commcare.formplayer.postgresutil.PostgresDB;
 import org.javarosa.core.services.PropertyManager;
 import org.javarosa.core.services.properties.Property;
 import org.javarosa.core.services.storage.IStorageIndexedFactory;
